feat: 新增basic auth授权方式

+++ git-checkout-core/src/main/kotlin/com/tencent/bk/devops/git/core/service/helper/auth/CredentialStoreAuthHelper.kt
@@ -104,11 +104,18 @@ class CredentialStoreAuthHelper(
     }
 
     override fun removeAuth() {
-        val storeCredentialValue = git.tryConfigGet(configKey = GitConstants.GIT_CREDENTIAL_HELPER)
-        val credentialFilePath = storeCredentialValue.substringAfter("--file=")
-            .removePrefix("'").removeSuffix("'")
-        Files.deleteIfExists(Paths.get(credentialFilePath))
-        git.tryConfigUnset(configKey = GitConstants.GIT_CREDENTIAL_HELPER)
+        val storeCredentialValue = git.tryConfigGet(
+            configKey = GitConstants.GIT_CREDENTIAL_HELPER,
+            configValueRegex = "store"
+        )
+        if (storeCredentialValue.isNotEmpty()) {
+            val credentialFilePath = storeCredentialValue.substringAfter("--file=")
+                .removePrefix("'").removeSuffix("'")
+            if (credentialFilePath.isNotEmpty()) {
+                Files.deleteIfExists(Paths.get(credentialFilePath))
+            }
+            git.tryConfigUnset(configKey = GitConstants.GIT_CREDENTIAL_HELPER)
+        }
     }
 
     override fun configGlobalAuth(copyGlobalConfig: Boolean) {
